more cleanup and restructuring
