+++ src/Bindings/RabbitMQAsyncCollector.cs
@@ -14,9 +14,9 @@ namespace Microsoft.Azure.WebJobs.Extensions.RabbitMQ
         private readonly RabbitMQContext _context;
         private readonly IBasicPublishBatch _batch;
 
-        private readonly ILogger<RabbitMQAsyncCollector> _logger;
+        private readonly ILogger _logger;
 
-        public RabbitMQAsyncCollector(RabbitMQContext context, ILogger<RabbitMQAsyncCollector> logger)
+        public RabbitMQAsyncCollector(RabbitMQContext context, ILogger logger)
         {
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
             _context = context ?? throw new ArgumentNullException(nameof(context));
@@ -31,7 +31,6 @@ namespace Microsoft.Azure.WebJobs.Extensions.RabbitMQ
         public Task AddAsync(byte[] message, CancellationToken cancellationToken = default)
         {
             _batch.Add(exchange: _context.ResolvedAttribute.Exchange, routingKey: _context.ResolvedAttribute.QueueName, mandatory: false, properties: _context.ResolvedAttribute.Properties, body: message);
-            _logger.LogDebug($"Message: {message}, Queue Name: {_context.ResolvedAttribute.QueueName}");
             _logger.LogDebug($"Adding message to batch for publishing...");
 
             return Task.CompletedTask;
