+++ test/WebJobs.Extensions.RabbitMQ.Samples/RabbitMQSamples.cs
@@ -1,11 +1,9 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using System.Json;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.RabbitMQ;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using RabbitMQ.Client.Events;
@@ -14,10 +12,11 @@ namespace WebJobs.Extensions.RabbitMQ.Samples
 {
     public static class RabbitMQSamples
     {
+        // Output samples
         public static void TimerTrigger_StringOutput(
             [TimerTrigger("00:01")] TimerInfo timer,
             [RabbitMQ(
-                Hostname = "localhost",
+                HostName = "localhost",
                 QueueName = "queue")] out string outputMessage,
             ILogger logger)
         {
@@ -28,7 +27,7 @@ namespace WebJobs.Extensions.RabbitMQ.Samples
         public static void TimerTrigger_PocoOutput(
              [TimerTrigger("00:01")] TimerInfo timer,
              [RabbitMQ(
-                  Hostname = "localhost",
+                  HostName = "localhost",
                   QueueName = "queue")] out TestClass outputMessage,
              ILogger logger)
         {
@@ -36,19 +35,18 @@ namespace WebJobs.Extensions.RabbitMQ.Samples
             logger.LogInformation($"RabbitMQ output binding message: {JsonConvert.SerializeObject(outputMessage)}");
         }
 
-        //// To run:
-        //// 1. Create Azure Storage Account and go to the homepage for that account
-        //// 2. Look for Queue service and click Queues on the sidebar
-        //// 3. Create a queue named "samples-rabbitmq-messages"
-        //// 4. Add a message to the queue
-        //// 5. Run this sample and you will see the queue trigger fired.
-        //// *Note that any time the queue isn't empty, the trigger will continue to fire.
-        //// So you can add items to the queue while the sample is running, and the trigger will be called until the queue is empty.
-
+        // To run:
+        // 1. Create Azure Storage Account and go to the homepage for that account
+        // 2. Look for Queue service and click Queues on the sidebar
+        // 3. Create a queue named "samples-rabbitmq-messages"
+        // 4. Add a message to the queue
+        // 5. Run this sample and you will see the queue trigger fired.
+        // *Note that any time the queue isn't empty, the trigger will continue to fire.
+        // So you can add items to the queue while the sample is running, and the trigger will be called until the queue is empty.
         public static async Task ProcessMessage_RabbitMQAsyncCollector(
             [QueueTrigger(@"samples-rabbitmq-messages")] string message,
             [RabbitMQ(
-                Hostname = "localhost",
+                HostName = "localhost",
                 QueueName = "queue")] IAsyncCollector<byte[]> messages,
             ILogger logger)
         {
@@ -57,19 +55,18 @@ namespace WebJobs.Extensions.RabbitMQ.Samples
             await messages.AddAsync(messageInBytes);
         }
 
-        //// To run:
-        //// 1. Create Azure Storage Account and go to the homepage for that account
-        //// 2. Look for Queue service and click Queues on the sidebar
-        //// 3. Create a queue named "samples-rabbitmq-messages"
-        //// 4. Add a message to the queue in POCO format (i.e.: "{ "name": Katie }")
-        //// 5. Run this sample and you will see the queue trigger fired.
-        //// *Note that any time the queue isn't empty, the trigger will continue to fire.
-        //// So you can add items to the queue while the sample is running, and the trigger will be called until the queue is empty.
-
+        // To run:
+        // 1. Create Azure Storage Account and go to the homepage for that account
+        // 2. Look for Queue service and click Queues on the sidebar
+        // 3. Create a queue named "samples-rabbitmq-messages"
+        // 4. Add a message to the queue in POCO format (i.e.: "{ "name": Katie }")
+        // 5. Run this sample and you will see the queue trigger fired.
+        // *Note that any time the queue isn't empty, the trigger will continue to fire.
+        // So you can add items to the queue while the sample is running, and the trigger will be called until the queue is empty.
         public static void QueueTrigger_RabbitMQOutput(
             [QueueTrigger(@"samples-rabbitmq-messages")] TestClass message,
             [RabbitMQ(
-                Hostname = "localhost",
+                HostName = "localhost",
                 QueueName = "queue")] out TestClass outputMessage,
             ILogger logger)
         {
@@ -94,18 +91,22 @@ namespace WebJobs.Extensions.RabbitMQ.Samples
             logger.LogInformation($"RabbitMQ queue trigger function processed message: {Encoding.UTF8.GetString(args.Body)}");
         }
 
-        public static void RabbitMQTrigger_Json(
-            [RabbitMQTrigger("localhost", "queue")] JsonValue jsonObj,
-            ILogger logger)
-        {
-            logger.LogInformation($"RabbitMQ queue trigger function processed message: {jsonObj}");
-        }
-
         public static void RabbitMQTrigger_JsonToPOCO(
             [RabbitMQTrigger("localhost", "queue")] TestClass pocObj,
             ILogger logger)
         {
             logger.LogInformation($"RabbitMQ queue trigger function processed message: {pocObj}");
+        }
+
+        public static void RabbitMQTrigger_RabbitMQOutput(
+            [RabbitMQTrigger("localhost", "queue")] string inputMessage,
+            [RabbitMQ(
+                HostName = "localhost",
+                QueueName = "hello")] out string outputMessage,
+            ILogger logger)
+        {
+            outputMessage = inputMessage;
+            logger.LogInformation($"RabittMQ output binding function sent message: {outputMessage}");
         }
 
         public class TestClass
